requirements.txt
fastapi
uvicorn[standard]
requests
playwright
pandas
numpy
matplotlib
beautifulsoup4
lxml
python-dotenv
PyPDF2
orjson
